Initial version
Simple Flask webapp for receiving images, processing them, and
displaying resulting images and some text. In addition, it hashes
every incoming image and caches the output of processing so that
the image and corresponding text can be served without doing the
processing from scratch. Currently the process() function is a stub
and cache location is hardcoded as temp directory.